minor update to import/export code. Added functionality to write empty file with numImgs in filename to export folder. Added code to remove all exported items if the overall export fails

+++ VMATTBICSIAutoPlanMT/ImportListener/Program.cs
@@ -156,7 +156,7 @@ namespace ImportListener
         {
             //increment the time on the progress window for each "tick", which is set to intervals of 1 second
             UpdateProgress();
-            elapsedSec += (double)(updateFrequencyMSec / 1000);
+            elapsedSec += updateFrequencyMSec / 1000;
         }
 
         private static void UpdateProgress()
